Betta version

+++ online_library/book_storage/views.py
@@ -6,10 +6,12 @@ from django.views.generic import ListView
 from .forms import CommentForm
 from .models import *
 from django.views.decorators.http import require_POST
+from django.db.models import Q
 
-menu = (('Жанры', 'home'), ('Авторы', 'authors_page'), ('Издательство', 'home'), ('О нас', 'about'))
+menu = (('Жанры', 'tags_page'), ('Авторы', 'authors_page'), ('О нас', 'about'))
 
-#class LibraryHome(ListView):
+
+# class LibraryHome(ListView):
 #    model = Book
 #    template_name = 'book_storage/index.html'
 #    context_object_name = 'books'
@@ -18,13 +20,13 @@ menu = (('Жанры', 'home'), ('Авторы', 'authors_page'), ('Издате
 
 def home(request, id_genre=None, type_of_book=None):
     if id_genre:
-        books = Book.objects.filter(id_genre=id_genre)
+        books = Book.objects.filter(id_genre=id_genre, is_published=True)
         message = f'Все книги с жанром: {Genre.objects.get(pk=id_genre)}'
     elif type_of_book:
-        books = Book.objects.filter(type_of_book=type_of_book)
+        books = Book.objects.filter(type_of_book=type_of_book, is_published=True)
         message = f'Все книги с тегом: {TypeOfBook.objects.get(pk=type_of_book)}'
     else:
-        books = Book.objects.all()
+        books = Book.objects.filter(is_published=True)
         message = ''
     context = {'title': 'Главная', 'menu': menu, 'books': books, 'message': message}
     return render(request, 'book_storage/index.html', context)
@@ -58,17 +60,20 @@ def book_page(request, id_book):
             'ser_books': ser_books, 'rate': rate, 'comments': comments, 'form': form, 'check': check}
     return render(request, 'book_storage/book_page.html', data)
 
+
 def author_page(request, id_author):
     author = Author.objects.get(pk=id_author)
     books = Book.objects.filter(id_author=id_author)
     data = {'title': f'Страница автора', 'author': author, 'books': books, 'menu': menu}
     return render(request, 'book_storage/author_page.html', data)
 
+
 def authors_page(request):
     authors = Author.objects.all()
     books = Book.objects.all()
     data = {'title': f'Все авторы', 'authors': authors, 'menu': menu}
     return render(request, 'book_storage/authors_page.html', data)
+
 
 @login_required
 def add_liked_book(request, id_book):
@@ -77,6 +82,7 @@ def add_liked_book(request, id_book):
     if not lk_book.exists():
         LikedBook.objects.create(user=request.user, liked_book=book)
     return HttpResponseRedirect(request.META['HTTP_REFERER'])
+
 
 @login_required
 def remove_liked_book(request, id_lk_book):
@@ -89,6 +95,7 @@ def read_page(request, id_book):
     book = Book.objects.get(pk=id_book)
     data = {'title': f'{book.title}', 'menu': menu, 'book': book}
     return render(request, 'book_storage/read_page.html', data)
+
 
 def about_page(request):
     context = {'title': 'Кто мы?', 'menu': menu}
@@ -135,9 +142,33 @@ def book_comments(request, id_book):
     context = {'form': form}
     return render(request, 'book_storage/comment.html', context)
 
+
+@login_required
 def delete_comment(request, id_comment):
     comment = Comment.objects.get(pk=id_comment)
     id_book = comment.book.id
     uri = reverse('book_page', args=(id_book,))
-    comment.delete()
+    if request.user == comment.user:
+        comment.delete()
     return redirect(uri)
+
+
+def search(request):
+    search_quary = request.GET.get('search-all', '')
+    if search_quary:
+        books = Book.objects.filter(Q(title__icontains=search_quary) | Q(description__icontains=search_quary),
+                                    is_published=True)
+        authors = Author.objects.filter(fio_author__icontains=search_quary)
+        #seria = BookSeries.objects.filter(book_series__icontains=search_quary)
+    else:
+        return HttpResponseRedirect(reverse('home'))
+
+    data = {'title': 'Результаты поиска', 'books': books, 'authors': authors, 'menu': menu}
+    return render(request, 'book_storage/search_page.html', data)
+
+
+def tags_page(request):
+    genres = Genre.objects.all()
+    types = TypeOfBook.objects.all()
+    data = {'title': 'Выбрать книгу', 'genres': genres, 'types': types, 'menu': menu}
+    return render(request, 'book_storage/tags_page.html', data)
